requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alpha-vantage==2.3.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.29.1
frozenlist==1.3.0
graphviz==0.19.1
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
multidict==6.0.2
neat-python==0.92
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pycodestyle==2.8.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
toml==0.10.2
urllib3==1.26.8
yarl==1.7.2
